Created Album view with tracks
- added requests for searching artists by name and getting info by album id
- created view for album view
- created second in the string duration converter
