+++ Musiclib/Store/Album/AlbumEnvironment.swift
@@ -1,0 +1,12 @@
+//
+//  AlbumEnvironment.swift
+//  Musiclib
+//
+//  Created by mlunts on 22.03.2022.
+//
+
+import ComposableArchitecture
+
+struct AlbumEnvironment {
+    var albumRequest: (_ albumId: Int) -> Effect<Album, APIError>
+}
