+++ Musiclib/Store/Album/AlbumState.swift
@@ -1,0 +1,18 @@
+//
+//  albumState.swift
+//  Musiclib
+//
+//  Created by mlunts on 22.03.2022.
+//
+
+struct AlbumState: Equatable {
+    var album: Album
+
+    var tracks: [Track] {
+        if let tracks = album.tracks?.data {
+            return tracks
+        } else { return [] }
+    }
+    
+    var isLoading: Bool = false
+}
