+++ Musiclib/Store/Artist/ArtistAction.swift
@@ -1,0 +1,13 @@
+//
+//  ArtistAction.swift
+//  musiclib
+//
+//  Created by mlunts on 21.03.2022.
+//
+
+import Foundation
+
+enum ArtistAction {
+    case onAppear
+    case albumsDataLoaded(Result<[Album], APIError>)
+}
