+++ Musiclib/Store/Artist/ArtistEnvironment.swift
@@ -1,0 +1,13 @@
+//
+//  ArtistEnvironment.swift
+//  musiclib
+//
+//  Created by mlunts on 21.03.2022.
+//
+
+import Foundation
+import ComposableArchitecture
+
+struct ArtistEnvironment {
+    var artistAlbumsRequest: (_ artistId: Int) -> Effect<[Album], APIError>
+}
