+++ Musiclib/Store/Artist/ArtistState.swift
@@ -1,0 +1,11 @@
+//
+//  ArtistState.swift
+//  musiclib
+//
+//  Created by mlunts on 21.03.2022.
+//
+
+struct ArtistState: Equatable {
+    var artist: Artist
+    var isLoading: Bool = false
+}
