--- musiclib/Store/Artist/ArtistState.swift
@@ -1,12 +1,0 @@
-//
-//  ArtistState.swift
-//  musiclib
-//
-//  Created by mlunts on 21.03.2022.
-//
-
-import Foundation
-struct ArtistState: Equatable {
-    var artist: Artist
-    var isLoading: Bool = false
-}
