--- musiclib/Store/Charts/ChartsAction.swift
@@ -1,13 +1,0 @@
-//
-//  ChartsAction.swift
-//  musiclib
-//
-//  Created by mlunts on 21.03.2022.
-//
-
-import Foundation
-
-enum ChartsAction {
-    case onAppear
-    case dataLoaded(Result<Chart, APIError>)
-}
