--- musiclib/Store/Charts/ChartsState.swift
@@ -1,13 +1,0 @@
-//
-//  ChartsState.swift
-//  musiclib
-//
-//  Created by mlunts on 21.03.2022.
-//
-
-import Foundation
-
-struct ChartsState: Equatable {
-    var artists: [Artist]?
-}
-
